third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/nonces/import-enforce-blocked-expected.txt

CONSOLE ERROR: line 21: Refused to load the script 'http://127.0.0.1:8000/security/resources/blank.html' because it violates the following Content Security Policy directive: "script-src 'nonce-abc'".

This is a testharness.js-based test.
PASS Incorrectly nonced imports do not load. 
PASS Incorrectly nonced imports trigger reports. 
Harness: the test ran to completion.
